test: split test file on components for each command
